% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.plot.R
\docType{package}
\name{coda.plot}
\alias{coda.plot-package}
\alias{coda.plot}
\title{coda.plot}
\description{
Easy to run function to built visualisations by means of 'ggplot2' objects.
}
\author{
Marc Comas-Cufí
}
