% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_diagram.R
\name{add_ternary_grid}
\alias{add_ternary_grid}
\title{Add a ternary grid layer}
\usage{
add_ternary_grid(p, ticks = seq(0.1, 0.9, 0.1), n = 300, eps = 1e-06, ...)
}
\arguments{
\item{p}{A \code{ggplot2} object created by \code{ternary_base()}.}

\item{ticks}{Numeric vector of grid levels.}

\item{n}{Number of sampled points per grid line.}

\item{eps}{Small positive offset used before log-ratio transformation.}

\item{...}{Further arguments passed to \code{ggplot2::geom_path()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Add a ternary grid layer
}
