% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{balance_dendrogram}
\alias{balance_dendrogram}
\title{Compositional balance dendrogram}
\usage{
balance_dendrogram(X, B)
}
\arguments{
\item{X}{Compositional data set}

\item{B}{Balance matrix}
}
\value{
ggplot2 object with the balance dendrogram
}
\description{
Compositional balance dendrogram
}
