% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_ciplot.R
\name{clr_biplot}
\alias{clr_biplot}
\title{Compositional clr-biplot}
\usage{
clr_biplot(X, alpha = 1, col_group = NULL, shape_group = NULL)
}
\arguments{
\item{X}{Compositional data}

\item{alpha}{Numeric values between 0 and 1, default alpha=1 (0: covariance biplot, 1: form biplot)}

\item{col_group}{Factor used to colorise observations}

\item{shape_group}{Factor used to shape observations}
}
\value{
a ggplot object with the biplot
}
\description{
Compositional clr-biplot
}
