% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean_barplot.R
\name{geometric_mean_barplot}
\alias{geometric_mean_barplot}
\title{Geometric Mean Barplot}
\usage{
geometric_mean_barplot(
  X,
  group,
  group_label = "",
  title = "Geometric mean bar plot",
  xlab = "",
  ylab = "ln(group mean)-ln(overall mean)",
  facets = TRUE,
  x_show_parts = TRUE
)
}
\arguments{
\item{X}{Composition Matrix}

\item{group}{Groups to compare}

\item{group_label}{Label for groups}

\item{title}{Plot title}

\item{xlab}{X label}

\item{ylab}{Y label}

\item{facets}{Should the barplot be separated by facets.}

\item{x_show_parts}{Should the barplots show the relation within parts.}
}
\description{
Geometric Mean Barplot
}
