% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean_barplot.R
\name{geometric_mean_barplot}
\alias{geometric_mean_barplot}
\title{Geometric Mean Barplot}
\usage{
geometric_mean_barplot(
  X,
  group,
  group_label = "",
  title = "Geometric mean bar plot",
  xlab = "",
  ylab = "ln(group gmean)-ln(overall gmean)",
  x_show_parts = TRUE,
  clr_scale = FALSE
)
}
\arguments{
\item{X}{A matrix or data frame representing compositional data.}

\item{group}{An optional factor indicating groups for comparison.}

\item{group_label}{A character string used as the label for the grouping variable in the legend.}

\item{title}{A character string specifying the plot title.}

\item{xlab}{A character string specifying the label for the x-axis.}

\item{ylab}{A character string specifying the label for the y-axis.}

\item{x_show_parts}{Logical. If \code{TRUE}, the x-axis shows parts to visualize the relative contributions within the composition.}

\item{clr_scale}{Should the clr coordinates be compared?}
}
\value{
A \code{ggplot} object representing the barplot.
}
\description{
Creates a barplot of the geometric mean of compositional parts.
}
