% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary}
\alias{ternary}
\title{Ternary diagram}
\usage{
ternary(X, groups = NULL)
}
\arguments{
\item{X}{Compositional data}
}
\value{
a ggtern object with the ternary diagram
}
\description{
Ternary diagram
}
