% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary}
\alias{ternary}
\title{Ternary diagram}
\usage{
ternary(X, group = NULL, center = FALSE, scale = FALSE)
}
\arguments{
\item{X}{Compositional data}

\item{center}{Should the data be centered?}

\item{scale}{Should the data be scaled to have unit variance?}

\item{groups}{Colorize according to this vector}
}
\value{
a ggtern object with the ternary diagram
}
\description{
Ternary diagram
}
